feat: add simulation controllers

+++ server/src/main/java/com/hyperlogix/server/services/simulation/SimulationStatus.java
@@ -4,5 +4,5 @@ package com.hyperlogix.server.services.simulation;
 public record SimulationStatus(
     boolean running,
     boolean paused,
-    int timeAcceleration) {
+    double timeAcceleration) {
 }
